==> notebooks/01_data_cleaning_netflix_titles.ipynb <==
# Limpieza del dataframe netflix_titles.csv
## Junio 2025
## by Francis

### Importar funciones generales

import sys
sys.path.append(r"C:\Users\FRANCIS\Documents\Proyecto_Streaming\scripts") 
import utils
import pandas as pd

import importlib
import utils
importlib.reload(utils)

### Carga del DataFrame

df_netflix = utils.cargar_datos(r"C:\Users\FRANCIS\Documents\Proyecto_Streaming\data\netflix_titles.csv")

### Inspección del DataFrame

df_netflix.head()

df_netflix.info()

### Convertir columna con fechas a datetime 

df_netflix = utils.convertir_a_fecha(df_netflix, ['date_added'])

df_netflix.dtypes

### Ver porcentaje de valores nulos

utils.resumen_nulos(df_netflix)

### Evaluamos en las columnas que no son importantes para los objetivos de los análisis y que tienen valores nulos 

evaluar_director = utils.evaluar_fila(df_netflix, "director", ["date_added", "rating", "show_id","listed_in", "country", "type"])
evaluar_director.to_string() #sí hay suficientes nulos para eliminar filas (70%< es not null)

df_netflix = df_netflix.drop(evaluar_director.index)

evaluar_cast = utils.evaluar_fila(df_netflix, "cast", ["date_added", "rating", "show_id", "listed_in", "country", "type"])
evaluar_cast.to_string() #no hay suficientes nulos para eliminar filas (>70% es not null)

evaluar_duration = utils.evaluar_fila(df_netflix, "duration", ["date_added", "rating", "show_id", "listed_in", "country", "type"])
evaluar_duration.to_string() #no hay suficientes nulos para eliminar filas (>70% es not null)

### Reemplazamos en todas las columnas object del DataFrame los datos nulos por "Desconocido"

columnas_object = ["director", "country", "rating", "duration", "cast"]
df_netflix[columnas_object] = df_netflix[columnas_object].fillna("Desconocido")

### Se convierte date_added a int para poder calcular la moda y mediana dentro de sus datos

df_netflix["date_added"] = df_netflix["date_added"].astype("int64")


df_netflix.dtypes

### Se aplica la función que reemplaza los valores nulos en date_added por la mediana o moda de la agrupación de datos 

df_netflix = utils.elegir_imputacion(df_netflix, columna="date_added", agrupacion=["listed_in", "type", "country"], umbral_dispersion=60)

### Se verifica si hay valores extraños o valores nulos después del reemplazo

print(df_netflix["date_added"].unique())  
print(df_netflix["date_added"].isnull().sum())  

### Los valores están en formato int con nanosegundos: se hace la conversión a segundos con timestamp para convertir a datetime

# Filtrar valores extremos antes de la conversión
df_netflix = df_netflix[df_netflix["date_added"] // 10**9 > 0]  
df_netflix = df_netflix[df_netflix["date_added"] // 10**9 < 1893456000]  

df_netflix["date_added"] = pd.to_datetime(df_netflix["date_added"] // 10**9, unit="s", errors="coerce")


### Se verifica aleatoriamente los datos, se confirma si hay valores nulos todavía y si la conversión a datetime fue realizada

print(df_netflix[["date_added"]].sample(10))  
print(df_netflix["date_added"].isnull().sum())  
df_netflix.dtypes


### Se crean las columnas derivadas de date_added y se convierten a tipo int

df_netflix["month_added"] = df_netflix["date_added"].dt.month.astype("Int64")  
df_netflix["year_added"] = df_netflix["date_added"].dt.year.astype("Int64")

print(df_netflix["date_added"].isnull().sum())  # ¿Quedan valores nulos después de todo?
print(df_netflix[["date_added", "month_added", "year_added"]].sample(10))  # Revisión rápida

utils.resumen_nulos(df_netflix)

### Datos duplicados

df_netflix.duplicated().sum() #NO HAY DATOS DUPLICADOS 

### Estandarizar los datos según la columna

df_netflix = utils.limpiar_texto(df_netflix, ["type", "country", "rating", "listed_in","title", "director", "cast","duration", "description"], conservar_mayusculas=["title", "director", "cast","duration", "description"]) 

df_netflix.head() 

df_netflix.dtypes

### Guardar el DataFrame limpio

df_netflix.to_csv("netflix_limpio.csv", index=False)  
df_netflix.to_parquet("netflix_limpio.parquet")
