==> notebooks/03_data_cleaning_disney_plus_tittles.ipynb <==
# Limpieza del dataframe disney_plus_titles.csv
## Junio 2025
## by Francis

### Importar funciones generales

import sys
sys.path.append(r"C:\Users\FRANCIS\Documents\Proyecto_Streaming\scripts") 
import utils
import pandas as pd

### Carga del Dataframe

df_disney = utils.cargar_datos(r"C:\Users\FRANCIS\Documents\Proyecto_Streaming\data\disney_plus_titles.csv")

### Inspección del DataFrame

df_disney.head()

df_disney.info()

### Convertir columna con fechas a datetime

df_disney= utils.convertir_a_fecha(df_disney, ['date_added'])

df_disney.dtypes

### Ver porcentaje de valores nulos

utils.resumen_nulos(df_disney)

### Evaluamos en las columnas que no son importantes para los objetivos de los análisis y que tienen valores nulos 

evaluar_director = utils.evaluar_fila(df_disney, "director", ["date_added", "rating", "show_id","listed_in", "country", "type"])
evaluar_director.to_string() #sí hay suficientes nulos para eliminar filas (70%< es not null)

df_disney = df_disney.drop(evaluar_director.index)

evaluar_cast = utils.evaluar_fila(df_disney, "cast", ["date_added", "rating", "show_id", "listed_in", "country", "type"])
evaluar_cast.to_string() #no hay suficientes nulos para eliminar filas (>70% es not null)

### Reemplazamos en todas las columnas object del DataFrame los datos nulos por "Desconocido"

columnas_object= ['country', 'rating',"director", "cast"]
df_disney[columnas_object]= df_disney[columnas_object].fillna("Desconocido")

### Se convierte date_added a int para poder calcular la moda y mediana dentro de sus datos

df_disney['date_added']= df_disney['date_added'].astype('int64')

df_disney.dtypes

### Se aplica la función que reemplaza los valores nulos en date_added por la mediana o moda de la agrupación de datos 

df_disney = utils.elegir_imputacion(df_disney, columna="date_added", agrupacion=["listed_in", "type", "country"], umbral_dispersion=60)

### Se verifica si hay valores extraños o valores nulos después del reemplazo

print(df_disney["date_added"].unique())  
print(df_disney["date_added"].isnull().sum())  

### Los valores están en formato int con nanosegundos: se hace la conversión a segundos con timestamp para convertir a datetime

# Filtrar valores extremos antes de la conversión
df_disney = df_disney[df_disney["date_added"] // 10**9 > 0]  
df_disney = df_disney[df_disney["date_added"] // 10**9 < 1893456000] 

df_disney["date_added"] = pd.to_datetime(df_disney["date_added"] // 10**9, unit="s", errors="coerce")


### Se verifica aleatoriamente los datos, se confirma si hay valores nulos todavía y si la conversión a datetime fue realizada

print(df_disney[['date_added']].sample(10))
print(df_disney['date_added'].isnull().sum())
df_disney.dtypes

### Se crean las columnas derivadas de date_added y se convierten a tipo int

df_disney['year_added']= df_disney['date_added'].dt.year.astype("Int64")
df_disney['month_added']= df_disney['date_added'].dt.month.astype("Int64")

print(df_disney["date_added"].isnull().sum())  # ¿Quedan valores nulos después de todo?
print(df_disney[["date_added", "month_added", "year_added"]].sample(10))  # Revisión rápida

utils.resumen_nulos(df_disney)

### Datos duplicados

df_disney.duplicated().sum() #No hay datos duplicados

### Estandarizar los datos según la columna

df_disney = utils.limpiar_texto(df_disney, ["type", "country", "rating", "listed_in","title", "director", "cast","duration", "description"], conservar_mayusculas=["title", "director", "cast","duration", "description"]) 

df_disney.head()

df_disney.dtypes

### Guardar el DataFrame limpio

df_disney.to_csv("disney_limpio.csv", index=False)  
df_disney.to_parquet("disney_limpio.parquet")
