==> notebooks/05_data_analysis.ipynb <==
### Cargar los DataFrames limpios

import pandas as pd

df_netflix = pd.read_parquet("netflix_limpio.parquet")

df_amazon = pd.read_parquet("amazon_limpio.parquet")

df_disney = pd.read_parquet("disney_limpio.parquet")

### Concatenar los DataFrames 

df_disney["plataforma"] = "Disney"
df_netflix["plataforma"] = "Netflix"
df_amazon["plataforma"] = "Amazon"

df_streaming = pd.concat([df_disney, df_netflix, df_amazon], ignore_index=True)

df_streaming.to_parquet("df_streaming_completo.parquet")

df_streaming.tail()

df_streaming["plataforma"].value_counts()

## Análisis

### Top 5 géneros más populares por plataforma

# Asegura que 'listed_in' sea una lista
df_streaming["listed_in"] = df_streaming["listed_in"].apply(
    lambda x: [i.strip() for i in x.split(",")] if isinstance(x, str) else x
)

# Explota los géneros
df_genero = df_streaming.explode("listed_in")
df_genero['listed_in'] = df_genero['listed_in'].str.strip()

# Agrupa por plataforma y género
genero_plataforma = df_genero.groupby(["plataforma", "listed_in"]).size().reset_index(name="conteo")

genero_plataforma = genero_plataforma.sort_values(by=["plataforma", "conteo"], ascending=[True, False])

# Mostrar los 5 géneros más populares de cada plataforma
top5_por_plataforma = genero_plataforma.groupby("plataforma").head(5).reset_index(drop=True)
top5_por_plataforma 

top5_por_plataforma.to_parquet("top5_por_plataforma.parquet")

### ¿Cuántas películas y series se añaden a la plataforma cada año?

df_anual = df_streaming.groupby(["year_added", "type","plataforma"]).size().reset_index(name="conteo")
df_anual.tail()

df_anual.to_parquet("df_anual.parquet")

### Top 5 años más productivos para cada plataforma

# Agrupa y ordena por plataforma y conteo descendente
df_productivos = df_streaming.groupby(["plataforma", "year_added"]).size().reset_index(name="conteo")
df_productivos = df_productivos.sort_values(["plataforma", "conteo"], ascending=[True, False])

# Selecciona el top 5 años más productivos para cada plataforma
top5_productivos = df_productivos.groupby("plataforma").head(5).reset_index(drop=True)

top5_productivos

top5_productivos.to_parquet("top5_productivos.parquet")

df = pd.read_parquet("df_streaming_completo.parquet")
anio_mas_antiguo_amazon = df[df["plataforma"] == "Amazon"]["year_added"].max()
print("El año más reciente en que se añadió un contenido en Amazon es:", anio_mas_antiguo_amazon)

df = pd.read_parquet("df_streaming_completo.parquet")
anio_mas_antiguo_amazon = df[df["plataforma"] == "Amazon"]["year_added"].min()
print("El año más antiguo en que se añadió un contenido en Amazon es:", anio_mas_antiguo_amazon)

### ¿Son más populares las películas o series en los últimos 5 años?

# Filtra los últimos 5 años
df_reciente = df_streaming[df_streaming["year_added"] >= 2018]

# Agrupa por plataforma y tipo
popularidad_reciente = (
    df_reciente.groupby(["plataforma", "type"])
    .size()
    .reset_index(name="conteo")
    .sort_values(["plataforma", "conteo"], ascending=[True, False])
)

popularidad_reciente

popularidad_reciente.to_parquet("popularidad_reciente.parquet")

### Top 5 países con más contenido por plataforma

# Asegura de que country sea lista
df_streaming["country"] = df_streaming["country"].apply(
    lambda x: [i.strip() for i in x.split(",")] if isinstance(x, str) else x
)

# Explota la columna country
df_paises = df_streaming.explode("country")
df_paises["country"] = df_paises["country"].str.strip()

# Agrupa por plataforma y país, cuenta títulos
conteo_paises = df_paises.groupby(["plataforma", "country"]).size().reset_index(name="conteo")

# Ordena y selecciona el top 5 países por plataforma
top5_paises_por_plataforma = (
    conteo_paises.sort_values(["plataforma", "conteo"], ascending=[True, False])
    .groupby("plataforma")
    .head(5)
    .reset_index(drop=True)
)
top5_paises_por_plataforma

top5_paises_por_plataforma.to_parquet("top5_paises_por_plataforma.parquet")

### Top 5 rating del contenido por plataforma

# Agrupa por plataforma y clasificación, cuenta títulos
df_rating_plat = df_streaming.groupby(["plataforma", "rating"]).size().reset_index(name="conteo")

# Ordena y selecciona el top 5 clasificaciones por plataforma
top5_rating_por_plataforma = (
    df_rating_plat.sort_values(["plataforma", "conteo"], ascending=[True, False])
    .groupby("plataforma")
    .head(5)
    .reset_index(drop=True)
)
top5_rating_por_plataforma

top5_rating_por_plataforma.to_parquet("top5_rating_por_plataforma.parquet")

### Top 5 clasificaciones por tipo de contenido

# Agrupa por tipo y clasificación, cuenta títulos
df_rating_tipo = df_streaming.groupby(["type", "rating"]).size().reset_index(name="conteo")

# Ordena y selecciona el top 5 clasificaciones por tipo de contenido
top5_rating_por_tipo = (
    df_rating_tipo.sort_values(["type", "conteo"], ascending=[True, False])
    .groupby("type")
    .head(5)
    .reset_index(drop=True)
)
top5_rating_por_tipo

top5_rating_por_tipo.to_parquet("top5_rating_por_tipo.parquet")
